06-lab-complete/requirements.txt
fastapi==0.115.12
google-genai==1.70.0
openai==2.30.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.0.1
redis==5.2.1
uvicorn==0.34.0
